--- Dockefile
@@ -1,8 +1,0 @@
-FROM golang
-WORKDIR /root
-
-ADD main.go .
-
-EXPOSE 4000
-
-ENTRYPOINT ["go", "run", "main.go"]
