--- README.md
@@ -1,9 +1,0 @@
-# 2019_2_LeMMaS
-
-Project "Agario"
-
-### Developers
-<a name="developers"></a>
-
-- [Sudarev Maxim](https://github.com/smi97)
-- [Mozhevikina Leonarda](https://github.com/ledka17)
