+++ src/Dockefile
@@ -1,0 +1,8 @@
+FROM golang
+WORKDIR /root
+
+ADD main.go .
+
+EXPOSE 4000
+
+ENTRYPOINT ["go", "run", "main.go"]
